Move some code to parametric scenario API, reuse tested image to get library version
